+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+- Add support for negative numbers.
+
 ## [v1.1.1] - 2017-05-16
 - Fix `--non-zero-exit-on-violation` option.